First commit of Robot control GUI.
Currently the GUI can start a ROS node and receive current and goal
position/velocity for each servo. It also enables the creation of a list
of joint poses from the current robot’s state, and the ability to add them
to a second list, which will act as a queue for moving through sequences
of poses later on. The lists can be manipulated and also saved/loaded to
CSV text files. A group of slider widgets shows the robot’s current joint
positions, with added markers to show ‘target’ positions.